refactored into sendgrid component
